requirements.txt
# Titan MCP Server - Python Dependencies

# MCP (Model Context Protocol)
mcp>=1.0.0

# LLM Providers
openai>=1.0.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Local Embeddings (CPU-only, no OpenAI dependency)
fastembed>=0.2.0

# Discord Bot
discord.py>=2.3.0

# Airtable
pyairtable>=2.2.0

# Google Drive API (for Connections.csv monitoring)
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Logging & Monitoring
structlog>=24.1.0
apify-client>=1.6.0
